+++ CHANGELOG.md
@@ -10,7 +10,9 @@ Note that since we don't clearly distinguish between a public and private interf
     - Add ``timing-mode`` Viewer GET param
     - Add support for webgl timer queries
     - Add timer marks around GPU render & compute operations
+- Volume Server CIF: Add check that a data block contains volume data before parsing
 - Fix ``Scene.clear`` not clearing primitives & volumes arrays (@JonStargaryen)
+- Fix rendering volumes when wboit is switched off and postprocessing is enabled
 
 ## [v3.8.2] - 2022-05-22